chore: refactor to move endpage block to generic folder (#7248)
* chore: refactor to move endpage block to generic folder

* chore: add thank you and payment ty svgr images to generic folder
