+++ frontend/src/features/public-form/components/FormEndPage/components/PaymentEndPagePreview.tsx
@@ -4,10 +4,10 @@ import { FormColorTheme, FormDto } from '~shared/types/form'
 
 import { SubmissionData } from '~features/public-form/PublicFormContext'
 
+import { PaymentsThankYouSvgr } from '../../../../../components/FormEndPage/PaymentsThankYouSvgr'
 import { PaymentEndPageBlock } from '../../../../admin-form/create/end-page/PaymentEndPageBlock'
 
 import { FeedbackBlock, FeedbackFormInput } from './FeedbackBlock'
-import { PaymentsThankYouSvgr } from './PaymentsThankYouSvgr'
 
 export interface PaymentEndPagePreviewProps {
   endPage: FormDto['endPage']
